+++ src/expand/join_dollar.c
@@ -6,7 +6,7 @@
 /*   By: ertrigna <ertrigna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/13 14:00:24 by vdeliere          #+#    #+#             */
-/*   Updated: 2025/06/19 16:38:29 by ertrigna         ###   ########.fr       */
+/*   Updated: 2025/06/30 09:18:24 by ertrigna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
