+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: ertrigna <ertrigna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/13 10:14:20 by ertrigna          #+#    #+#             */
-/*   Updated: 2025/06/25 17:02:13 by ertrigna         ###   ########.fr       */
+/*   Updated: 2025/06/30 09:48:05 by ertrigna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
